+++ lib/core/constants/images.dart
@@ -1,0 +1,15 @@
+class AppImages{
+  static const logo = 'assets/images/logo.png';
+  static const offer = 'assets/images/offer.png';
+  static const food = 'assets/images/food.png';
+  static const food2 = 'assets/images/food2.png';
+  static const food3 = 'assets/images/food3.png';
+  static const food4 = 'assets/images/food4.png';
+  static const food5 = 'assets/images/food5.png';
+  static const food6 = 'assets/images/food6.png';
+  static const restaurant = 'assets/images/restaurant.jpg';
+  static const restaurant2 = 'assets/images/restaurant2.jpg';
+  static const restaurant3 = 'assets/images/restaurant3.jpg';
+  static const restaurant4 = 'assets/images/restaurant4.jpg';
+  static const restaurant5 = 'assets/images/restaurant5.jpg';
+}